[feat] add support for outside-of-transaction migrations (#269)
Some PostgreSQL commands cannot be used inside transactions.

This change introduces the ability to do PostgreSQL migrations outside
of a transaction. When defining a transaction with `Generate()` or
`Computed()`, transactional or non-transactional will be determined by
the type signature of the callback function.

For `Script()`, the text will be examined and a determination made based
on command matching. To force it one way or the other, use the
`libschema.ForceTransactional()` or `libschema.ForceNonTransactional()`
migration options.

MySQL and SingleStore packages were updated to support the same pattern,
though they already did most things non-transactionally because they
don't support any DDL inside transactions.

Also changed in this PR:

- Switched from google.com/pkg/errors to github.com/memsql/errors
- Use sentinel errors instead for key situations